src/live_monitoring/requirements.txt
python packages:
customTKinter
PILLOW